refactor(test): fix naming and add comments
