+++ packages/integration-tests/src/ui-helpers/expect-page.ts
@@ -2,6 +2,7 @@ import { type ElementHandle, type Page } from 'puppeteer';
 
 import { expectNavigation } from '#src/utils.js';
 
+/** Error thrown by {@link ExpectPage}. */
 class ExpectPageError extends Error {
   constructor(
     message: string,
@@ -17,20 +18,43 @@ class ExpectPageError extends Error {
 export default class ExpectPage {
   constructor(public readonly page = global.page) {}
 
+  /**
+   * Navigate to the given URL and wait for the page to load. Assert that an element with ID `app`
+   * is present.
+   *
+   * @param initialUrl The URL to navigate to.
+   */
   async toStart(initialUrl: URL) {
     await expectNavigation(this.page.goto(initialUrl.href), this.page);
     await expect(this.page).toMatchElement('#app');
   }
 
+  /**
+   * Click on the element matching the given selector and text.
+   *
+   * @param selector The selector to match.
+   * @param text The text to match, if provided.
+   * @param shouldNavigate Whether the click should trigger a navigation. Defaults to `true`.
+   */
   async toClick(selector: string, text?: string | RegExp, shouldNavigate = true) {
     const clicked = expect(this.page).toClick(selector, { text });
     return shouldNavigate ? expectNavigation(clicked, this.page) : clicked;
   }
 
+  /**
+   * Click on the `<button type="submit">` element on the page.
+   *
+   * @param shouldNavigate Whether the click should trigger a navigation. Defaults to `true`.
+   */
   async toClickSubmit(shouldNavigate = true) {
     return this.toClick('button[type=submit]', undefined, shouldNavigate);
   }
 
+  /**
+   * Find the `<form>` element on the page and submit it programmatically.
+   *
+   * @param shouldNavigate Whether the click should trigger a navigation. Defaults to `true`.
+   */
   async toSubmit(shouldNavigate = true) {
     const form = await expect(this.page).toMatchElement('form');
     // eslint-disable-next-line no-restricted-syntax
@@ -40,6 +64,16 @@ export default class ExpectPage {
     return shouldNavigate ? expectNavigation(submitted, this.page) : submitted;
   }
 
+  /**
+   * Fill an `<input>` with the given name with the given value and optionally submit the form.
+   *
+   * @param name The name of the input to fill.
+   * @param value The value to fill the input with.
+   * @param options Options to control the behavior of the method.
+   * @param options.submit Whether to submit the form after filling the input. Defaults to `false`.
+   * @param options.shouldNavigate Whether the submit should trigger a navigation. Defaults to `true`. Note that this
+   * option is ignored if `options.submit` is `false`.
+   */
   async toFillInput(
     name: string,
     value: string,
@@ -51,6 +85,15 @@ export default class ExpectPage {
     }
   }
 
+  /**
+   * Fill a `<form>` with the given values and optionally submit it.
+   *
+   * @param values The key-value object of values to fill the form with.
+   * @param options Options to control the behavior of the method.
+   * @param options.submit Whether to submit the form after filling the input. Defaults to `false`.
+   * @param options.shouldNavigate Whether the submit should trigger a navigation. Defaults to `true`. Note that this
+   * option is ignored if `options.submit` is `false`.
+   */
   async toFillForm(
     values: Record<string, string>,
     options?: { submit: true; shouldNavigate?: boolean }
@@ -61,14 +104,29 @@ export default class ExpectPage {
     }
   }
 
+  /**
+   * Expect the page to match an element with `role="alert"` and optionally with the given text.
+   *
+   * @param text The text to match, if provided.
+   */
   async toMatchAlert(text?: string | RegExp): Promise<ElementHandle> {
     return expect(this.page).toMatchElement('*[role=alert]', { text });
   }
 
+  /**
+   * Expect the page's URL to match the given URL.
+   *
+   * @param url The URL to match.
+   */
   toMatchUrl(url: URL | string) {
     expect(this.page.url()).toBe(typeof url === 'string' ? url : url.href);
   }
 
+  /**
+   * Expect a toast to appear with the given text, then remove it immediately.
+   *
+   * @param text The text to match.
+   */
   async waitForToast(text: string | RegExp) {
     const toast = await expect(this.page).toMatchElement(`.ReactModal__Content[class*=toast]`, {
       text,
